requirements.txt
amqp==2.6.1
arrow==0.17.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.2
django-timezone-field==4.0
djangorestframework==3.12.1
idna==2.10
kombu==4.6.11
pkg-resources==0.0.0
prompt-toolkit==3.0.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
